Fix: Removed bugs in hash, insert, find, and topk/bottomk
In hash function, the reduction calculation had errors that made the value negative. Fixed by preserving size_t value until returning the final value
Insert was not using reference which caused existing items to not be updated.
Fixed find where it didn't have anything indicating the key was not found.
Top and bottom k were not functioning properly. Made the resize only occur if the size of the vector was more than k, and also used references in the for loop.